requirements.txt
# Web Scraping
selenium==4.15.2
webdriver-manager==4.0.1

# Data Processing
pandas>=2.2.0  # Python 3.13 compatible
numpy>=1.26.0

# Language Detection
langdetect==1.0.9

# Text Analysis
nltk>=3.8.1
textblob==0.17.1
vaderSentiment==3.3.2

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
wordcloud>=1.9.0

# Configuration
PyYAML>=6.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.27.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
emoji>=2.4.0